# docs/swagger.yaml
definitions:
...
  link.PayByLinkRequest:
    properties:
      pay_key:
        maxLength: 6
        type: string
      remark:
        maxLength: 100
        type: string
      token:
        type: string
    required:
    - pay_key
    - token
    type: object
  link.PaymentLinkRequest:
    properties:
      amount:
        type: number
      product_name:
        maxLength: 30
        type: string
      remark:
        maxLength: 100
        type: string
      total_limit:
        minimum: 1
        type: integer
      user_limit:
        minimum: 1
        type: integer
    required:
    - amount
    - product_name
    type: object
...
  util.ResponseAny:
    properties:
      data: {}
      error_msg:
        example: ""
        type: string
    type: object
info:
  contact: {}
  title: LINUX DO Credit
  version: 1.0.0
paths:
...
  /api/v1/merchant/api-keys/{id}/payment-links:
    get:
      parameters:
      - description: API Key ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - merchant
    post:
      consumes:
      - application/json
      parameters:
      - description: API Key ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: 创建支付链接请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/link.PaymentLinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - merchant
  /api/v1/merchant/api-keys/{id}/payment-links/{linkId}:
    delete:
      parameters:
      - description: API Key ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: Payment Link ID
        format: int64
        in: path
        name: linkId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - merchant
    put:
      consumes:
      - application/json
      parameters:
      - description: API Key ID
        format: int64
        in: path
        name: id
        required: true
        type: integer
      - description: Payment Link ID
        format: int64
        in: path
        name: linkId
        required: true
        type: integer
      - description: 更新支付链接请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/link.PaymentLinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - merchant
...
  /api/v1/merchant/payment-links/{token}:
    get:
      parameters:
      - description: 支付链接 Token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - merchant
  /api/v1/merchant/payment-links/pay:
    post:
      consumes:
      - application/json
      parameters:
      - description: 支付请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/link.PayByLinkRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - merchant
...
swagger: "2.0"